==> analysis/analyses/4-method_characterisation/1a_process_methods.R <==
# This file will process the method characteristics google sheets
library(tidyverse)
library(googlesheets)
library(dynalysis)

experiment("method_characteristics")

# Downloading -----------------------
# # If it's your first time running this script, run this:
# gs_auth()

methods <- gs_key("1Mug0yz8BebzWt8cmEW306ie645SBh_tDHwjVw4OFhlE") %>%
  gs_read(ws = "Software", col_types = cols(GScholarClusterID = "c"), skip = 1)

# Dates ------------------------------
methods$date <- methods$Preprint
methods$date[is.na(methods$date)] <- methods$PubDate[is.na(methods$date)]

# Num citations ---------------------------------------
# Don't run this in parallel ─ you can't rush this
# methods$ncitations <- pbapply::pbsapply(methods$GScholarClusterID, dynutils::google_scholar_num_citations)
library(rcrossref)
methods$ncitations <- pbapply::pbsapply(cl=4, methods$DOI, cr_citation_count)

# Trajectory components --------------------------
# split maximal trajectory types
methods <- methods %>%
  mutate(maximal_trajectory_types_split = map(maximal_trajectory_types, ~as.character(stringr::str_split(., "[ ]?,[ ]?", simplify=TRUE)))) %>%
  mutate(maximal_trajectory_type = map_chr(maximal_trajectory_types_split, first))

# now add for every trajectory type a column, whether it can handle such a trajectory or not
trajectory_type_capabilities <- map(
  methods$maximal_trajectory_types_split,
  function(maximal_trajectory_types) {
    ancestors <- trajectory_types %>%
      filter(id %in% maximal_trajectory_types) %>%
      pull(ancestors) %>%
      unlist() %>%
      unique()
    trajectory_types$id %in% ancestors
  }) %>%
  map(~as_tibble(as.list(setNames(., trajectory_types$id)))) %>%
  bind_rows()

methods <- methods %>% bind_cols(trajectory_type_capabilities)

## Non inclusion reasons ------------------------
methods$non_inclusion_reasons_split <- methods$non_inclusion_reasons %>% str_split("[ ]?,[ ]?")

## Method conversions ------------------------
methods$conversion_split <- methods$conversion %>% str_split("[ ]?,[ ]?")
allowed_conversions <- c("branching_local", "linear", "branching_cluster", "branching_global", "branching_local_projection", "special", NA)
if(!methods$conversion_split %>% map_lgl(~all(. %in% allowed_conversions)) %>% all()) {
  stop("Some conversions are invalid!")
}
methods$conversion_special <- map_lgl(methods$conversion_split, ~"special" %in% .)

## Method outputs --------------
methods$output_split <- methods$output %>% str_split("[ ]?,[ ]?")


# add extra reason for date cutoff
date_cutoff <- as.Date("2017-05-01")
date_filter <- methods$date > date_cutoff
methods$non_inclusion_reasons_split[date_filter] <- map(methods$non_inclusion_reasons_split[date_filter], c, "date") %>% map(unique)

# Saving -------------------------
write_rds(methods, derived_file("methods.rds"))
