==> analysis/analyses/4-method_characterisation/2_quality_control.Rmd <==
---
title: "Method quality control"
output:
  html_document:
    df_print: paged
---

```{r message=F}
library(tidyverse)
library(dynalysis)
library(cowplot)

experiment("method_characteristics")

method_qc <- readRDS(derived_file("method_qc.rds"))
methods_evaluated <- readRDS(derived_file("methods_evaluated.rds"))
checks <- readRDS(derived_file("checks.rds"))
```

Table for the different checks
```{r, results="asis"}
library(kableExtra)
table <- checks %>% 
  mutate(
    weight = cell_spec(
      pritt("<span style='display: none;'>{aspect_id}</span> {weight}"), # add an invisible grouping indicator
      background = spec_color(weight, end=0.9,option="A",direction=-1),
      escape=FALSE,
      color="white"
    )
  ) %>%
  mutate(
    references = references %>% 
      gsub(",", "; ", .) %>% 
      {ifelse(!is.na(.), pritt("[{.}]"), "")} %>%
      {pritt("<span style='display: none;'>{aspect_id}</span> {.}")} # add an invisible grouping indicator
  ) %>% 
  select(aspect, weight, references, item, item_weight, !!qc_applications$application) %>% 
  mutate_at(vars(!!qc_applications$application), ~ifelse(., "✓", "")) %>% 
  rename_all(label_long) %>% 
  knitr::kable("html", escape=FALSE) %>% 
  kable_styling(bootstrap_options = "striped") %>% 
  collapse_rows(columns=1:3)

for (category in unique(checks$category)) {
  color <- qc_categories %>% filter(category == !!category) %>% pull(color)
  table <- table %>% 
    group_rows(
      label_long(category), 
      min(which(checks$category == category)), 
      max(which(checks$category == category)),
      label_row_css = pritt("background-color: {color}; color: #fff;")
  )
}
  
table %>% write_rds(figure_file("qc_checks.rds"))
```


Comparing qc and citations overtime     
```{r, fig.width=16, fig.height=8}
g1 <- ggplot(methods_evaluated, aes(date, qc_score)) +
    geom_smooth(span=2) +
    geom_point() +
    ggrepel::geom_label_repel(aes(label = name)) +
    cowplot::theme_cowplot() +
    labs(x = label_long("publication_date"), y = "QC score", title = "Code quality score over time")
g1

g2 <- ggplot(methods_evaluated %>% mutate(ncitations = ifelse(ncitations == 0, 0.5, ncitations)), aes(date, ncitations)) +
    geom_smooth(span=2) +
    geom_point() +
    ggrepel::geom_label_repel(aes(label = name)) +
    cowplot::theme_cowplot() +
    scale_y_log10() +
    labs(x = "Time", y = "Citations", title = "Citations over time")
g2

g3 <- ggplot(methods_evaluated %>% mutate(ncitations = ifelse(ncitations == 0, 0.5, ncitations)), aes(ncitations, qc_score)) +
    geom_smooth(span=2) +
    geom_point() +
    ggrepel::geom_label_repel(aes(label = name)) +
    cowplot::theme_cowplot() +
    scale_x_log10() +
    labs(x = "Citations", y = "QC score", title = "QC score over # citations")
g3

cowplot::plot_grid(g1, g3, g2, ncol=2, nrow=2, labels = "auto") %>% 
  write_rds(figure_file("qc_over_time.rds"))
```


Calculate the overall qc scores for each method
```{r}
# set the ordering of method_ids to this ordering
method_ordering_plot <- methods_evaluated %>%
  arrange(-qc_score) %>% 
  mutate(method_id = factor(name, levels=name)) %>% 
  ggplot(aes(method_id, qc_score)) +
    geom_bar(aes(fill=qc_score), stat="identity") +
    theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
    scale_y_continuous(expand=c(0, 0), limits=c(0, 1)) +
    viridis::scale_fill_viridis(option="A", limits=c(0, 1))

method_ordering_plot
```



```{r, fig.width=10, fig.height=12}
qc_checks <- method_qc %>%
  filter(method_id == "SCORPIUS") %>%
  group_by(category) %>% 
  mutate(normalised_weight = weight/sum(weight))

aspect_position <- qc_checks %>% 
  arrange(aspect_id) %>%
  group_by(aspect_id) %>%
  summarise(
    aspect_width = sum(item_weight * normalised_weight),
    name = first(name)
  ) %>%
  mutate(
    aspect_end = cumsum(aspect_width),
    aspect_start = lag(aspect_end, 1, default=0),
    aspect_mid = aspect_start + (aspect_end - aspect_start)/2
  )

check_position <- qc_checks %>%
  arrange(aspect_id, check_id) %>%
  group_by(check_id) %>%
  summarise(
    check_width = item_weight * normalised_weight,
    aspect_id = first(aspect_id)
  ) %>%
  left_join(aspect_position, by="aspect_id") %>%
  group_by(aspect_id) %>%
  mutate(
    check_start = aspect_start + lag(cumsum(check_width), default=0),
    check_end = aspect_start + cumsum(check_width)
  )


check_heatmap_plot <- method_qc %>%
  left_join(check_position, by="check_id") %>%
  mutate(method_id = factor(method_id)) %>%
  mutate(answer = ifelse(is.na(answer), 0, answer)) %>%
  ggplot() +
    geom_rect(aes(xmax=check_start, xmin=check_end, ymin=as.integer(method_id), ymax=as.integer(method_id) + 1, fill=category, alpha=answer)) +
    geom_vline(aes(xintercept=check_end), data=check_position, color="#444444", alpha=0.25) +
    geom_vline(aes(xintercept=aspect_end), data=aspect_position, color="#444444") +
    scale_x_reverse("Check", breaks = aspect_position$aspect_mid, labels = aspect_position$name, expand = c(0, 0)) +
    scale_y_continuous("Method", breaks = seq_along(levels(method_qc$method_id))+0.5, labels = levels(method_qc$method_id), expand = c(0, 0)) +
    scale_fill_manual(values=set_names(qc_categories$color, qc_categories$category)) +
    coord_flip() +
    theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

check_heatmap_plot
```

Combined plot
```{r, fig.width=10, fig.height=12}
empty_x <- theme(
  axis.title.x=element_blank(),
  axis.text.x=element_blank(),
  axis.ticks.x=element_blank()
)

method_qc_check_plot <- cowplot::plot_grid(method_ordering_plot + empty_x, check_heatmap_plot, nrow=2, align = "v", axis = "rl", rel_heights = c(0.2, 0.9))

method_qc_check_plot %>% saveRDS(figure_file("method_qc_check.rds"))
```


## Order checks

Order by breakpoint (when ordering all methods by score)
```{r}
breakpoint <- function(x) {
  map_dbl(0:length(x), function(i) {
    sum(1-x[seq(1, i, length.out=i)]) +
      sum(x[seq(i, length(x), length.out=length(x)-i)])
  }) %>% which.min()
}
check_breakpoints <- method_qc %>%
  reshape2::acast(check_id~method_id, value.var="answer") %>%
  {
    tibble(
      breakpoint=apply(., 1, breakpoint),
      check_id = as.integer(rownames(.)),
      n_bad_methods = apply(., 1, sum)
    )
  } %>% 
  left_join(checks, by="check_id")
```

```{r fig.width=20}
method_qc %>% 
  mutate(answer = round(answer, 2)) %>% 
  group_by(check_id) %>% 
  count(answer) %>% 
  mutate(check_score = sum(answer * n)) %>% 
  ungroup() %>% 
  left_join(checks, by="check_id") %>% 
  arrange(-check_score) %>% 
  mutate(check_id = factor(check_id, levels=unique(check_id))) %>% 
  ggplot() +
    geom_bar(aes(check_id, n, fill=answer), stat="identity") +
    coord_flip() +
    scale_y_reverse(expand=c(0, 0)) +
    scale_x_discrete(labels=setNames(method_qc$item_weight, method_qc$check_id)) +
    scale_fill_distiller(palette="Blues", direction = 1)
```

```{r}
aspect_info <- method_qc %>% 
  group_by(aspect_id) %>% 
  summarise(
    name = first(name),
    category = first(category)
  )

aspect_ordering_plotdata <- method_qc %>%
  group_by(method_id, aspect_id) %>%
  summarise(
    qc_score=sum(answer * item_weight * weight)/sum(item_weight * weight)
  ) %>%
  ungroup() %>% 
  group_by(aspect_id) %>% 
  count(qc_score) %>% 
  mutate(aspect_ordering = sum((qc_score < 1) * n)) %>% 
  ungroup() %>% 
  arrange(aspect_ordering) %>% 
  left_join(aspect_info, by="aspect_id") %>% 
  mutate(aspect_id = factor(aspect_id, levels=unique(aspect_id)))

aspect_ordering_plotdata2 <- aspect_ordering_plotdata %>% 
  mutate(category = factor(category, levels=qc_categories$category)) %>% 
  arrange(-as.numeric(category), aspect_ordering) %>% 
  mutate(aspect_id = factor(aspect_id, levels=as.character(unique(aspect_id))))

aspect_ordering_plotdata2_category <- aspect_ordering_plotdata2 %>% 
  group_by(category, aspect_id) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(position = n()-seq_len(n())) %>% 
  group_by(category) %>% 
  summarise(start = min(position)+0.5, end=max(position)+1.5)

aspect_ordering_plot2 <- aspect_ordering_plotdata2 %>% 
  arrange(qc_score) %>% 
  ggplot() +
    geom_bar(aes(aspect_id, n, fill=set_names(qc_categories$color, qc_categories$category)[category], alpha=qc_score), stat="identity") +
    geom_vline(aes(xintercept=start), data=aspect_ordering_plotdata2_category) +
    scale_fill_identity() +
    scale_alpha_continuous(range=c(1, 0.05), name = "Quality control score", guide = guide_legend(title.position = "top", title.hjust = 0.5,label.position = "bottom"))
aspect_ordering_plot2

category_legend <- aspect_ordering_plotdata2 %>% 
  group_by(category, aspect_id) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(position = seq_len(n())) %>% 
  group_by(category) %>% 
  summarise(start = -min(position), end=-max(position)-1) %>% 
  ggplot() +
    geom_rect(aes(xmin=0, xmax=1, ymin=start, ymax=end), fill="white") +
    geom_text(aes(1, start + (end-start)/2, label=label_long(category), color=category), hjust=1, size=5) +
    geom_segment(aes(x=1.1, xend=1.1, y=start-0.2, yend=end+0.2, color=category)) +
    scale_color_manual(values=set_names(qc_categories$color, qc_categories$category)) +
    scale_y_continuous(expand=c(0, 0)) +
    theme_void() +
    theme(legend.position="none")
category_legend

n_methods <- length(unique(method_qc$method_id))

aspect_ordering_plot2 <- aspect_ordering_plot2 +
    coord_flip() +
    scale_y_continuous(expand=c(0, 0), name = "# methods", c(seq(0, n_methods, 5), n_methods)) +
    scale_x_discrete(labels=setNames(aspect_ordering_plotdata$name, aspect_ordering_plotdata$aspect_id), name="") +
    theme(legend.position="top")

method_qc_ordering_plot <- cowplot::plot_grid(category_legend, aspect_ordering_plot2, align = "h", rel_widths = c(0.2, 0.8), axis="tb")
method_qc_ordering_plot
saveRDS(method_qc_ordering_plot, figure_file("method_qc_ordering.rds"))
```
